[Framework] Move ArgumentList into TextureFilter interface
